+++ ingest_templates.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+# Ingest Templates
+cp -r $HOME/.config/safe_pocket/templates/. $HOME/DEV/bin/safe_pocket/src/defaults